requirements.txt
colorama
Pillow